feat: move StartedGame to /games/[id]/play
